+++ responsive_data_grid/lib/src/definitions/columns/duration.dart
@@ -12,6 +12,8 @@ class DurationColumn<TItem extends Object> extends GridColumn<TItem, Duration> {
     super.width,
     super.minWidth,
     super.maxWidth,
+    super.visible,
+    super.frozen,
     super.xlCols,
     super.largeCols,
     super.mediumCols,