+++ responsive_data_grid/lib/src/definitions/columns/enum.dart
@@ -16,6 +16,8 @@ class EnumColumn<TItem extends Object, TValue extends Enum>
     super.width,
     super.minWidth,
     super.maxWidth,
+    super.visible,
+    super.frozen,
     super.xlCols,
     super.largeCols,
     super.mediumCols,