+++ responsive_data_grid/lib/src/definitions/columns/num.dart
@@ -13,6 +13,8 @@ class NumColumn<TItem extends Object> extends GridColumn<TItem, num> {
     super.width,
     super.minWidth,
     super.maxWidth,
+    super.visible,
+    super.frozen,
     super.xlCols,
     super.largeCols,
     super.mediumCols,