+++ responsive_data_grid/lib/src/definitions/columns/string.dart
@@ -12,6 +12,8 @@ class StringColumn<TItem extends Object> extends GridColumn<TItem, String> {
     super.width,
     super.minWidth,
     super.maxWidth,
+    super.visible,
+    super.frozen,
     super.xlCols,
     super.largeCols,
     super.mediumCols,