+++ responsive_data_grid/lib/src/widgets/infinite_scroll_body.dart
@@ -86,17 +86,19 @@ class _ResponsiveGridInfiniteScrollBodyWidgetState<TItem extends Object>
               ),
               shrinkWrap: false,
               scrollDirection: Axis.vertical,
-              padding: widget.gridState.widget.padding.copyWith(
-                top: 0,
-                bottom: 0,
-              ),
+              padding: widget.gridState.widget.layoutMode == GridLayoutMode.table
+                  ? EdgeInsets.zero
+                  : widget.gridState.widget.padding.copyWith(
+                      top: 0,
+                      bottom: 0,
+                    ),
               builderDelegate: PagedChildBuilderDelegate(
                 noItemsFoundIndicatorBuilder: (context) =>
                     gridNoRecordsBody(widget.gridState),
                 itemBuilder: (context, item, index) {
                   return DataGridRowWidget<TItem>(
                     item: item,
-                    columns: widget.gridState.widget.columns,
+                    columns: widget.gridState.layoutColumns,
                     itemTapped: widget.gridState.widget.itemTapped,
                     theme: widget.theme,
                     padding: widget.gridState.widget.contentPadding,